requirements.txt
bidict==0.22.0
click==8.1.3
Flask==2.2.2
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-engineio==4.3.4
python-socketio==5.7.1
SQLAlchemy==1.4.40
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.8.1
